requirements.txt
Pillow==4.3.0
scipy==1.0.0rc1
numpy==1.12
scikit-learn==0.19
matplotlib==2.1